# src/main/java/com/bank/bootcoinservice/dto/associateyanki/AssociateYankiRequest.java
package com.bank.bootcoinservice.dto.associateyanki;

import lombok.Data;
import lombok.Getter;
import lombok.Setter;

@Data
@Getter
@Setter
public class AssociateYankiRequest {
    private String documentNumber;
}